impr: Move value lookup fun
