+++ flux-node/boot/init.go
@@ -2,7 +2,7 @@ package boot
 
 import (
 	"github.com/bytepowered/flux/flux-node"
-	"github.com/bytepowered/flux/flux-node/backend"
+	"github.com/bytepowered/flux/flux-node/common"
 	"github.com/bytepowered/flux/flux-node/discovery"
 	"github.com/bytepowered/flux/flux-node/ext"
 	"github.com/bytepowered/flux/flux-node/logger"
@@ -12,7 +12,7 @@ func init() {
 	// Default logger factory
 	ext.SetLoggerFactory(logger.DefaultFactory)
 	// 参数查找与解析函数
-	ext.SetArgumentLookupFunc(backend.DefaultArgumentLookupFunc)
+	ext.SetArgumentLookupFunc(common.LookupMTValue)
 	// Serializer
 	// Default: JSON
 	serializer := flux.NewJsonSerializer()
