+++ flux-node/common/resolver_mtval.go
@@ -1,4 +1,4 @@
-package backend
+package common
 
 import (
 	"bytes"
