+++ flux-node/common/resolver_mtval_test.go
@@ -1,4 +1,4 @@
-package backend
+package common
 
 import (
 	"encoding/json"
