+++ flux-node/testable/argument_test.go
@@ -2,7 +2,7 @@ package testable
 
 import (
 	"github.com/bytepowered/flux/flux-node"
-	"github.com/bytepowered/flux/flux-node/backend"
+	"github.com/bytepowered/flux/flux-node/common"
 	"github.com/bytepowered/flux/flux-node/context"
 	"github.com/bytepowered/flux/flux-node/ext"
 	assert2 "github.com/stretchr/testify/assert"
@@ -10,7 +10,7 @@ import (
 )
 
 func TestPrimitiveArgumentLookupResolve(t *testing.T) {
-	ext.SetArgumentLookupFunc(backend.DefaultArgumentLookupFunc)
+	ext.SetArgumentLookupFunc(common.LookupMTValue)
 	cases := []struct {
 		definition flux.Argument
 		class      string
@@ -122,7 +122,7 @@ func TestPrimitiveArgumentLookupResolve(t *testing.T) {
 }
 
 func TestComplexArgumentLookupResolve(t *testing.T) {
-	ext.SetArgumentLookupFunc(backend.DefaultArgumentLookupFunc)
+	ext.SetArgumentLookupFunc(common.LookupMTValue)
 	cases := []struct {
 		definition flux.Argument
 		class      string
